% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixopt.R
\name{mixsqp}
\alias{mixsqp}
\title{SQP algorithm with partial QR for optimizing mixtures.}
\usage{
mixsqp(L, x, convtol = 1e-08, pqrtol = 1e-08, eps = 1e-08,
  sptol = 0.001, maxiter = 100, maxqpiter = 100, seed = 1,
  verbose = TRUE)
}
\description{
Sequential quadratic programming (SQP) method with
    partial QR-based approximate computation of gradients and Hessians
    for efficiently solving the Mixture Distribution Optimization
    Problem.
}
